v1.2 (hotfix11)
Added a new (and hopefully *final*) sampler mode, where it will just run it for the duration of the note and then not loop again.
I needed that for so long and didn't have it :I
But ya might keep tweaking this but hopefully should be all good... I hope. Also going to update the manual with the sampler details to help out a bit. (for myself as well ._ .)